mock-server/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
email-validator==2.2.0
Flask==3.1.3
flask-cors==6.0.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.7
